lessons learned
Ember.js, from documentation
http://emberjs.com/guides/routing/specifying-a-routes-model/
A route with a dynamic segment will only have its model hook
called when it is entered via the URL. If the route is entered through a
transition (e.g. when using the link-to Handlebars helper), then a model
context is already provided and the hook is not executed. Routes without
dynamic segments will always execute the model hook.

active model serializers: be careful with serializer name
if you want custom serializer probably don't use "default" name
'system' attribute/name is used internally by gem, so it's better to not
use it